requirements.txt
setuptools
apache-airflow-providers-common-compat
apache-airflow-providers-apache-spark
pandas
delta-spark
deltalake
pyspark==3.5.5
pyarrow
faker
astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy